由温仰发起的RDS SDK自动发布, BUILD_ID=464, 版本号：2.1.2 发布日志： 1, Synchronize to the latest api list

+++ aliyun-net-sdk-rds/Rds/Transform/V20140815/DescribeSQLLogRecordsResponseUnmarshaller.cs
@@ -31,7 +31,7 @@ namespace Aliyun.Acs.Rds.Transform.V20140815
 
 			describeSQLLogRecordsResponse.HttpResponse = context.HttpResponse;
 			describeSQLLogRecordsResponse.RequestId = context.StringValue("DescribeSQLLogRecords.RequestId");
-			describeSQLLogRecordsResponse.TotalRecordCount = context.IntegerValue("DescribeSQLLogRecords.TotalRecordCount");
+			describeSQLLogRecordsResponse.TotalRecordCount = context.LongValue("DescribeSQLLogRecords.TotalRecordCount");
 			describeSQLLogRecordsResponse.PageNumber = context.IntegerValue("DescribeSQLLogRecords.PageNumber");
 			describeSQLLogRecordsResponse.PageRecordCount = context.IntegerValue("DescribeSQLLogRecords.PageRecordCount");
